document-qa-rag/backend/requirements.txt
fastapi
uvicorn
pypdf
chromadb
pymilvus
sentence-transformers
openai
langchain
tiktoken
python-dotenv